+++ src/main/java/codesquad/config/MainController.java
@@ -1,8 +1,6 @@
 package codesquad.config;
 
-import codesquad.question.QuestionRepository;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
